requirements.txt
absl-py==2.1.0
aiohttp==3.9.4
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.6.0
anyio==4.3.0
appscript==1.2.2
async==0.6.2
attrs==23.1.0
Automat==22.10.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
config==0.5.1
constantly==23.10.4
contourpy==1.1.0
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
distro==1.9.0
dm-tree==0.1.8
docutils==0.20.1
easygui==0.98.3
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Faker==19.4.0
ffmpy3==0.2.4
filelock==3.13.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.42.1
frozenlist==1.4.1
h11==0.14.0
h5py==3.10.0
hacker==0.3.1
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jmespath==1.0.1
keras==3.0.4
keyboard==0.13.5
keyring==24.2.0
kiwisolver==1.4.5
lxml==4.9.3
macholib==1.16.2
markdown-it-py==3.0.0
markup==0.2
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
modulegraph==0.19.5
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.5
namex==0.0.7
nh3==0.2.14
numpy==1.25.2
openai==0.28.0
openpyxl==3.1.2
outcome==1.2.0
package-name==0.1
packaging==23.1
pandas==2.1.0
parsel==1.8.1
Pillow==10.0.0
piwifi==0.3
pkginfo==1.9.6
platformdirs==3.11.0
playsound==1.3.0
plotly==5.17.0
pluggy==1.2.0
Protego==0.3.0
psutil==5.9.5
py2app==0.28.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
PyDispatcher==2.0.7
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.16.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.8
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc==10.0
pyobjc-core==10.0
pyobjc-framework-Accessibility==10.0
pyobjc-framework-Accounts==10.0
pyobjc-framework-AddressBook==10.0
pyobjc-framework-AdServices==10.0
pyobjc-framework-AdSupport==10.0
pyobjc-framework-AppleScriptKit==10.0
pyobjc-framework-AppleScriptObjC==10.0
pyobjc-framework-ApplicationServices==10.0
pyobjc-framework-AppTrackingTransparency==10.0
pyobjc-framework-AudioVideoBridging==10.0
pyobjc-framework-AuthenticationServices==10.0
pyobjc-framework-AutomaticAssessmentConfiguration==10.0
pyobjc-framework-Automator==10.0
pyobjc-framework-AVFoundation==10.0
pyobjc-framework-AVKit==10.0
pyobjc-framework-BusinessChat==10.0
pyobjc-framework-CalendarStore==10.0
pyobjc-framework-CallKit==10.0
pyobjc-framework-CFNetwork==10.0
pyobjc-framework-ClassKit==10.0
pyobjc-framework-CloudKit==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Collaboration==10.0
pyobjc-framework-ColorSync==10.0
pyobjc-framework-Contacts==10.0
pyobjc-framework-ContactsUI==10.0
pyobjc-framework-CoreAudio==10.0
pyobjc-framework-CoreAudioKit==10.0
pyobjc-framework-CoreBluetooth==10.0
pyobjc-framework-CoreData==10.0
pyobjc-framework-CoreHaptics==10.0
pyobjc-framework-CoreLocation==10.0
pyobjc-framework-CoreMedia==10.0
pyobjc-framework-CoreMediaIO==10.0
pyobjc-framework-CoreMIDI==10.0
pyobjc-framework-CoreML==10.0
pyobjc-framework-CoreMotion==10.0
pyobjc-framework-CoreServices==10.0
pyobjc-framework-CoreSpotlight==10.0
pyobjc-framework-CoreText==10.0
pyobjc-framework-CoreWLAN==10.0
pyobjc-framework-CryptoTokenKit==10.0
pyobjc-framework-DeviceCheck==10.0
pyobjc-framework-DictionaryServices==10.0
pyobjc-framework-DiscRecording==10.0
pyobjc-framework-DiscRecordingUI==10.0
pyobjc-framework-DiskArbitration==10.0
pyobjc-framework-DVDPlayback==10.0
pyobjc-framework-EventKit==10.0
pyobjc-framework-ExceptionHandling==10.0
pyobjc-framework-ExecutionPolicy==10.0
pyobjc-framework-ExternalAccessory==10.0
pyobjc-framework-FileProvider==10.0
pyobjc-framework-FileProviderUI==10.0
pyobjc-framework-FinderSync==10.0
pyobjc-framework-FSEvents==10.0
pyobjc-framework-GameCenter==10.0
pyobjc-framework-GameController==10.0
pyobjc-framework-GameKit==10.0
pyobjc-framework-GameplayKit==10.0
pyobjc-framework-ImageCaptureCore==10.0
pyobjc-framework-InputMethodKit==10.0
pyobjc-framework-InstallerPlugins==10.0
pyobjc-framework-InstantMessage==10.0
pyobjc-framework-Intents==10.0
pyobjc-framework-IOBluetooth==10.0
pyobjc-framework-IOBluetoothUI==10.0
pyobjc-framework-IOSurface==10.0
pyobjc-framework-iTunesLibrary==10.0
pyobjc-framework-KernelManagement==10.0
pyobjc-framework-LatentSemanticMapping==10.0
pyobjc-framework-LaunchServices==10.0
pyobjc-framework-libdispatch==10.0
pyobjc-framework-libxpc==10.0
pyobjc-framework-LinkPresentation==10.0
pyobjc-framework-LocalAuthentication==10.0
pyobjc-framework-MapKit==10.0
pyobjc-framework-MediaAccessibility==10.0
pyobjc-framework-MediaLibrary==10.0
pyobjc-framework-MediaPlayer==10.0
pyobjc-framework-MediaToolbox==10.0
pyobjc-framework-Metal==10.0
pyobjc-framework-MetalKit==10.0
pyobjc-framework-MetalPerformanceShaders==10.0
pyobjc-framework-MetalPerformanceShadersGraph==10.0
pyobjc-framework-MLCompute==10.0
pyobjc-framework-ModelIO==10.0
pyobjc-framework-MultipeerConnectivity==10.0
pyobjc-framework-NaturalLanguage==10.0
pyobjc-framework-NetFS==10.0
pyobjc-framework-Network==10.0
pyobjc-framework-NetworkExtension==10.0
pyobjc-framework-NotificationCenter==10.0
pyobjc-framework-OpenDirectory==10.0
pyobjc-framework-OSAKit==10.0
pyobjc-framework-OSLog==10.0
pyobjc-framework-PassKit==10.0
pyobjc-framework-PencilKit==10.0
pyobjc-framework-Photos==10.0
pyobjc-framework-PhotosUI==10.0
pyobjc-framework-PreferencePanes==10.0
pyobjc-framework-PushKit==10.0
pyobjc-framework-Quartz==10.0
pyobjc-framework-QuickLookThumbnailing==10.0
pyobjc-framework-ReplayKit==10.0
pyobjc-framework-SafariServices==10.0
pyobjc-framework-SceneKit==10.0
pyobjc-framework-ScreenSaver==10.0
pyobjc-framework-ScreenTime==10.0
pyobjc-framework-ScriptingBridge==10.0
pyobjc-framework-SearchKit==10.0
pyobjc-framework-Security==10.0
pyobjc-framework-SecurityFoundation==10.0
pyobjc-framework-SecurityInterface==10.0
pyobjc-framework-ServiceManagement==10.0
pyobjc-framework-Social==10.0
pyobjc-framework-SoundAnalysis==10.0
pyobjc-framework-Speech==10.0
pyobjc-framework-SpriteKit==10.0
pyobjc-framework-StoreKit==10.0
pyobjc-framework-SyncServices==10.0
pyobjc-framework-SystemConfiguration==10.0
pyobjc-framework-SystemExtensions==10.0
pyobjc-framework-UniformTypeIdentifiers==10.0
pyobjc-framework-UserNotifications==10.0
pyobjc-framework-UserNotificationsUI==10.0
pyobjc-framework-VideoSubscriberAccount==10.0
pyobjc-framework-VideoToolbox==10.0
pyobjc-framework-Virtualization==10.0
pyobjc-framework-Vision==10.0
pyobjc-framework-WebKit==10.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
pyquery==2.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PythonTurtle==0.3.2
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
pywifi==1.1.12
queuelib==1.6.2
readme-renderer==42.0
requests==2.31.0
requests-cache==1.1.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.3
rubicon-objc==0.4.6
Scrapy==2.11.0
selenium==4.12.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
somepackage==1.2.3
sortedcontainers==2.4.0
soupsieve==2.5
sprites==1.480
SQLAlchemy==2.0.28
sudo==1.0.0
tenacity==8.2.3
termcolor==2.4.0
tldextract==5.1.1
tqdm==4.66.2
trio==0.22.2
twine==4.0.2
Twisted==22.10.0
typing_extensions==4.8.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.4
w3lib==2.1.2
watchdog==4.0.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
wikipedia==1.4.0
wsproto==1.2.0
XLML==0.1.2
xlwings==0.30.12
yarl==1.9.4
you-get==0.4.1650
zipp==3.16.2
zope.interface==6.1
